refactor: move `enter_fullscreen` to `fullscreen` subfolder
